# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivityDraft": {
        "properties": {
          "actions_taken": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Actions Taken",
            "type": "string"
          },
          "commands_summary": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Commands Summary",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "end_datetime": {
            "maxLength": 120,
            "minLength": 1,
            "title": "End Datetime",
            "type": "string"
          },
          "root_cause": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Root Cause",
            "type": "string"
          },
          "start_datetime": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Start Datetime",
            "type": "string"
          },
          "summary": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "ticket_id": {
            "title": "Ticket Id",
            "type": "integer"
          },
          "validation_result": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Validation Result",
            "type": "string"
          }
        },
        "required": [
          "ticket_id",
          "start_datetime",
          "end_datetime",
          "summary",
          "root_cause",
          "actions_taken",
          "commands_summary",
          "validation_result"
        ],
        "title": "ActivityDraft",
        "type": "object"
      },
      "BackupNotApplicableCreate": {
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "recorded_by": {
            "default": "technician",
            "maxLength": 120,
            "minLength": 1,
            "title": "Recorded By",
            "type": "string"
          },
          "source_path": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          }
        },
        "required": [
          "reason"
        ],
        "title": "BackupNotApplicableCreate",
        "type": "object"
      },
      "BackupRecord": {
        "properties": {
          "backup_created": {
            "default": false,
            "title": "Backup Created",
            "type": "boolean"
          },
          "backup_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backup Path"
          },
          "backup_required": {
            "default": false,
            "title": "Backup Required",
            "type": "boolean"
          },
          "backup_type": {
            "title": "Backup Type",
            "type": "string"
          },
          "checksum_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum Before"
          },
          "command_execution_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Execution Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "group_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Before"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mode_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode Before"
          },
          "mtime_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mtime Before"
          },
          "owner_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Before"
          },
          "persistent_across_reboot": {
            "default": false,
            "title": "Persistent Across Reboot",
            "type": "boolean"
          },
          "post_change_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Change Hash"
          },
          "pre_change_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pre Change Hash"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "redacted": {
            "default": false,
            "title": "Redacted",
            "type": "boolean"
          },
          "restore_command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restore Command"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "sanitized_diff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sanitized Diff"
          },
          "size_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Before"
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          },
          "stored_content": {
            "default": false,
            "title": "Stored Content",
            "type": "boolean"
          },
          "ticket_id": {
            "title": "Ticket Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "run_id",
          "ticket_id",
          "backup_type",
          "reason",
          "created_at"
        ],
        "title": "BackupRecord",
        "type": "object"
      },
      "BackupRestoreProposalCreate": {
        "properties": {
          "proposed_by": {
            "default": "technician",
            "maxLength": 120,
            "minLength": 1,
            "title": "Proposed By",
            "type": "string"
          },
          "reason": {
            "default": "Restore from recorded rollback command.",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "BackupRestoreProposalCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntegrationRequest": {
        "properties": {
          "activity_draft_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Draft Id"
          },
          "activity_payload": {
            "additionalProperties": true,
            "title": "Activity Payload",
            "type": "object"
          },
          "attempts": {
            "default": 0,
            "title": "Attempts",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "phoenix_activity_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phoenix Activity Id"
          },
          "request_type": {
            "title": "Request Type",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/IntegrationRequestStatus"
          },
          "ticket_id": {
            "title": "Ticket Id",
            "type": "integer"
          },
          "ticket_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TicketStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "ticket_id",
          "request_type",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "IntegrationRequest",
        "type": "object"
      },
      "IntegrationRequestStatus": {
        "enum": [
          "pending",
          "processing",
          "activity_created",
          "completed",
          "failed",
          "dead_letter"
        ],
        "title": "IntegrationRequestStatus",
        "type": "string"
      },
...
      "ProposedStep": {
        "properties": {
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "approved_command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Command"
          },
          "command": {
            "title": "Command",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expected_signal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Signal"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "risk_class": {
            "title": "Risk Class",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "safety_notes": {
            "items": {
              "type": "string"
            },
            "title": "Safety Notes",
            "type": "array"
          },
          "safety_summary": {
            "title": "Safety Summary",
            "type": "string"
          },
          "safety_verdict": {
            "title": "Safety Verdict",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/StepStatus"
          },
          "timeout_s": {
            "title": "Timeout S",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "run_id",
          "created_at",
          "source",
          "phase",
          "command",
          "purpose",
          "risk_class",
          "safety_verdict",
          "safety_summary",
          "status",
          "timeout_s"
        ],
        "title": "ProposedStep",
        "type": "object"
      },
...
      "StepStatus": {
        "enum": [
          "proposed",
          "approved",
          "rejected",
          "blocked",
          "executed",
          "failed"
        ],
        "title": "StepStatus",
        "type": "string"
      },
...
      "TicketStatus": {
        "enum": [
          "OPEN",
          "PENDING",
          "DONE"
        ],
        "title": "TicketStatus",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend control plane for Phoenix ERP, technician approvals, SSH execution, and audit state.",
    "title": "techbold AI Service Desk Autopilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs/{run_id}/activity/draft": {
      "post": {
        "operationId": "draft_run_activity_api_runs__run_id__activity_draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Run Activity"
      },
      "put": {
        "operationId": "save_run_activity_draft_api_runs__run_id__activity_draft_put",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "edited_by",
            "required": false,
            "schema": {
              "default": "technician",
              "title": "Edited By",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Run Activity Draft"
      }
    },
    "/api/runs/{run_id}/activity/save": {
      "post": {
        "operationId": "save_run_activity_draft_api_runs__run_id__activity_save_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "edited_by",
            "required": false,
            "schema": {
              "default": "technician",
              "title": "Edited By",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Run Activity Draft"
      }
    },
    "/api/runs/{run_id}/activity/submit": {
      "post": {
        "operationId": "submit_run_activity_api_runs__run_id__activity_submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationRequest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Run Activity"
      }
    },
    "/api/runs/{run_id}/backups": {
      "get": {
        "operationId": "get_run_backups_api_runs__run_id__backups_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BackupRecord"
                  },
                  "title": "Response Get Run Backups Api Runs  Run Id  Backups Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Backups"
      }
    },
    "/api/runs/{run_id}/backups/not-applicable": {
      "post": {
        "operationId": "record_backup_not_applicable_api_runs__run_id__backups_not_applicable_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupNotApplicableCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Backup Not Applicable"
      }
    },
    "/api/runs/{run_id}/backups/{backup_record_id}/restore": {
      "post": {
        "operationId": "propose_backup_restore_api_runs__run_id__backups__backup_record_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "backup_record_id",
            "required": true,
            "schema": {
              "title": "Backup Record Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRestoreProposalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposedStep"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose Backup Restore"
      }
    },
...
  }
}